Déplacement de l'écran DossierProcédure dans l'onglet Instruction
