+++ scripts/front-end/components/screens/Dossier.svelte
@@ -3,12 +3,9 @@
 
     import Squelette from '../Squelette.svelte'
     //import TagPhase from '../TagPhase.svelte'
-    import TagEnjeu from '../TagEnjeu.svelte'
-    import EnteteDossier from './Dossier/EnteteDossier.svelte'
-    import DossierTabs from './Dossier/DossierTabs.svelte'
+    import EnteteDossier from '../Dossier/EnteteDossier.svelte'
+    import DossierTabs from '../Dossier/DossierTabs.svelte'
     
-    import {formatLocalisation, formatDéposant, phases, prochaineActionAttenduePar} from '../../affichageDossier.js'
-    import { modifierDossier } from '../../actions/dossier.js';
 
     /** @import {DossierComplet, DossierPhase} from '../../../types/API_Pitchou.ts' */
 
@@ -20,57 +17,6 @@
     /** @type {string | undefined} */
     export let email = undefined
     
-    $: phaseActuelle = dossier.évènementsPhase[0].phase;
-    
-    /** @type {Pick<DossierComplet, 'prochaine_action_attendue_par'> & {phase: DossierPhase}} */
-    let dossierParams = {
-        phase: phaseActuelle,
-        prochaine_action_attendue_par: dossier.prochaine_action_attendue_par,
-    }
-
-    $: dossierParams.phase = phaseActuelle
-
-
-    let messageErreur = "" 
-    let afficherMessageSucces = false
-
-    /**
-     * 
-     * @param {Event} e
-     */
-    const mettreAJourDossier = (e) => {
-        e.preventDefault()
-
-        /** @type {Partial<DossierComplet>} */
-        const modifs = {}
-
-        if(phaseActuelle !== dossierParams.phase){
-            modifs.évènementsPhase = [
-                {
-                    dossier: dossier.id,
-                    horodatage: new Date(),
-                    phase: dossierParams.phase,
-                    cause_personne: null
-                }
-            ]
-        }
-        
-        if(dossier.prochaine_action_attendue_par !== dossierParams.prochaine_action_attendue_par){
-            modifs.prochaine_action_attendue_par = dossierParams.prochaine_action_attendue_par
-        }
-
-        modifierDossier(dossier, modifs)
-            .then(() => afficherMessageSucces = true)
-            .catch((error) => {
-                console.info(error)
-                messageErreur = "Quelque chose s'est mal passé du côté serveur."
-            })
-    }
-
-    const retirerAlert = () => { 
-        messageErreur = ""
-        afficherMessageSucces = false
-    }
 </script>
 
 <Squelette {email}>
@@ -78,82 +24,6 @@
         <div class="fr-col">
             <EnteteDossier {dossier}></EnteteDossier>
             <DossierTabs {dossier} {messages}></DossierTabs>
-
-            <article class="fr-p-3w fr-mb-4w">
-                <section>
-                    <h2 class="fr-h5">Phase et prochaine action attendue</h2>
-                    
-                    <form class=" fr-mb-4w" on:submit={mettreAJourDossier} on:change={retirerAlert}>
-                        {#if messageErreur}
-                            <div class="fr-alert fr-alert--error fr-mb-3w">
-                                <h3 class="fr-alert__title">Erreur lors de la mise à jour :</h3>
-                                <p>{messageErreur}</p>
-                            </div>
-                        {/if}
-                        {#if afficherMessageSucces}
-                        <div class="fr-alert fr-alert--success fr-mb-3w">
-                            <p>La phase et de qui est attendu la prochaine action ont été mises à jour !</p>
-                        </div>
-                        {/if}
-                        <div class="fr-input-group">
-                            <label class="fr-label" for="phase">
-                                Phase du dossier
-                            </label>
-                    
-                            <select bind:value={dossierParams["phase"]} class="fr-select" id="phase">
-                                {#each [...phases] as phase}
-                                    <option value={phase}>{phase}</option>
-                                {/each}
-                            </select>
-                        </div>
-                        <div class="fr-input-group">
-                            <label class="fr-label" for="prochaine_action_attendue_par">
-                                Acteur(s) concerné(s)
-                            </label>
-                    
-                            <select bind:value={dossierParams["prochaine_action_attendue_par"]} class="fr-select" id="prochaine_action_attendue_par">
-                                {#each [...prochaineActionAttenduePar] as acteur}
-                                    <option value={acteur}>{acteur}</option>
-                                {/each}
-                            </select>
-                        </div>
-                        <button class="fr-btn" type="submit">
-                            Mettre à jour la phase ou de qui est attendu la prochaine action
-                        </button>
-                    </form>
-                </section>
-                <section>
-                    <h2 class="fr-h5">Informations</h2>
-                    <ul>
-                        <li>
-                            <strong>Porteur de projet</strong> : {formatDéposant(dossier)}<br />
-                        </li>
-                        <li>
-                            <strong>Localisation</strong> : {formatLocalisation(dossier)}
-                        </li>
-                        {#if dossier.espèces_protégées_concernées}
-                        <li class="liste-espèces">
-                            <strong>Liste des espèces protégées concernées</strong>
-                            <br>
-                            <a href={dossier.espèces_protégées_concernées}>{dossier.espèces_protégées_concernées}</a> :
-                        </li>
-                        {/if}
-
-                        {#if dossier.enjeu_politique || dossier.enjeu_écologique}
-                            <li>
-                                <strong>Enjeux</strong> : 
-                                {#if dossier.enjeu_politique}
-                                    <TagEnjeu enjeu="politique" taille='SM' classes={["fr-mb-1w"]}></TagEnjeu>
-                                {/if}
-
-                                {#if dossier.enjeu_écologique}
-                                    <TagEnjeu enjeu="écologique" taille='SM' classes={["fr-mb-1w"]}></TagEnjeu>
-                                {/if}
-                            </li>
-                        {/if}
-                    </ul>
-                </section>
-            </article>
         </div>
     </div>
 </Squelette>
