+++ scripts/front-end/main.js
@@ -7,7 +7,6 @@ import Dossier from './routes/Dossier.js';
 import RedactionArretePrefectoral from './routes/RedactionArretePrefectoral.js';
 import DossierMessagerie from './routes/DossierMessagerie.js'
 import DossierDescription from './routes/DossierDescription.js'
-import DossierProcédure from './routes/DossierProcédure.js'
 import SaisieEspèces from './routes/SaisieEspèces.js';
 import ImportHistoriqueNouvelleAquitaine from './routes/import-historique/NouvelleAquitaine.js'
 import PreremplissageDerogation from './routes/PreremplissageDerogation.js';
@@ -20,7 +19,6 @@ page('/', Accueil)
 
 page('/dossier/:dossierId', Dossier)
 page('/dossier/:dossierId/description', DossierDescription)
-page('/dossier/:dossierId/procedure', DossierProcédure)
 page('/dossier/:dossierId/messagerie', DossierMessagerie)
 page('/dossier/:dossierId/redaction-arrete-prefectoral', RedactionArretePrefectoral)
