+++ tsconfig.json
@@ -14,6 +14,7 @@
     "compilerOptions": {
       "allowJs": true,
       "checkJs": true,
+      "allowImportingTsExtensions": true,
       "noEmit": true,
       "module": "esnext",
       "strict": true,
@@ -24,7 +25,6 @@
       "downlevelIteration": true,
       "types": [],
       "noUnusedLocals": true,
-      "noUnusedParameters": true,
-      "allowImportingTsExtensions": true
+      "noUnusedParameters": true
     }
 }
